Moved chessboard and circles grid detectors to objdetect.
