+++ modules/calib/misc/java/test/CalibTest.java
@@ -18,86 +18,6 @@ import org.opencv.imgproc.Imgproc;
 
 public class CalibTest extends OpenCVTestCase {
 
-    Size size;
-
-    @Override
-    protected void setUp() throws Exception {
-        super.setUp();
-
-        size = new Size(3, 3);
-    }
-
-    public void testFindChessboardCornersMatSizeMat() {
-        Size patternSize = new Size(9, 6);
-        MatOfPoint2f corners = new MatOfPoint2f();
-        Calib.findChessboardCorners(grayChess, patternSize, corners);
-        assertFalse(corners.empty());
-    }
-
-    public void testFindChessboardCornersMatSizeMatInt() {
-        Size patternSize = new Size(9, 6);
-        MatOfPoint2f corners = new MatOfPoint2f();
-        Calib.findChessboardCorners(grayChess, patternSize, corners, Calib.CALIB_CB_ADAPTIVE_THRESH + Calib.CALIB_CB_NORMALIZE_IMAGE
-                + Calib.CALIB_CB_FAST_CHECK);
-        assertFalse(corners.empty());
-    }
-
-    public void testFind4QuadCornerSubpix() {
-        Size patternSize = new Size(9, 6);
-        MatOfPoint2f corners = new MatOfPoint2f();
-        Size region_size = new Size(5, 5);
-        Calib.findChessboardCorners(grayChess, patternSize, corners);
-        Calib.find4QuadCornerSubpix(grayChess, corners, region_size);
-        assertFalse(corners.empty());
-    }
-
-    public void testFindCirclesGridMatSizeMat() {
-        int size = 300;
-        Mat img = new Mat(size, size, CvType.CV_8U);
-        img.setTo(new Scalar(255));
-        Mat centers = new Mat();
-
-        assertFalse(Calib.findCirclesGrid(img, new Size(5, 5), centers));
-
-        for (int i = 0; i < 5; i++)
-            for (int j = 0; j < 5; j++) {
-                Point pt = new Point(size * (2 * i + 1) / 10, size * (2 * j + 1) / 10);
-                Imgproc.circle(img, pt, 10, new Scalar(0), -1);
-            }
-
-        assertTrue(Calib.findCirclesGrid(img, new Size(5, 5), centers));
-
-        assertEquals(1, centers.rows());
-        assertEquals(25, centers.cols());
-        assertEquals(CvType.CV_32FC2, centers.type());
-    }
-
-    public void testFindCirclesGridMatSizeMatInt() {
-        int size = 300;
-        Mat img = new Mat(size, size, CvType.CV_8U);
-        img.setTo(new Scalar(255));
-        Mat centers = new Mat();
-
-        assertFalse(Calib.findCirclesGrid(img, new Size(3, 5), centers, Calib.CALIB_CB_CLUSTERING
-                | Calib.CALIB_CB_ASYMMETRIC_GRID));
-
-        int step = size * 2 / 15;
-        int offsetx = size / 6;
-        int offsety = (size - 4 * step) / 2;
-        for (int i = 0; i < 3; i++)
-            for (int j = 0; j < 5; j++) {
-                Point pt = new Point(offsetx + (2 * i + j % 2) * step, offsety + step * j);
-                Imgproc.circle(img, pt, 10, new Scalar(0), -1);
-            }
-
-        assertTrue(Calib.findCirclesGrid(img, new Size(3, 5), centers, Calib.CALIB_CB_CLUSTERING
-                | Calib.CALIB_CB_ASYMMETRIC_GRID));
-
-        assertEquals(1, centers.rows());
-        assertEquals(15, centers.cols());
-        assertEquals(CvType.CV_32FC2, centers.type());
-    }
-
     public void testConstants()
     {
         // calib3d.hpp: some constants have conflict with constants from 'fisheye' namespace
