+++ modules/objdetect/CMakeLists.txt
@@ -2,6 +2,7 @@ set(the_description "Object Detection")
 ocv_define_module(objdetect
     opencv_core
     opencv_imgproc
+    opencv_features
     opencv_3d
     OPTIONAL
         opencv_dnn
