+++ modules/objdetect/misc/objc/gen_dict.json
@@ -3,5 +3,10 @@
         "QRCodeDetectorAruco": {
             "getDetectorParameters": { "declaration" : [""], "implementation" : [""] }
         }
+    },
+    "func_arg_fix" : {
+        "Calib" : {
+            "findCirclesGrid" : { "blobDetector" : {"defval" : "cv::SimpleBlobDetector::create()"} }
+        }
     }
 }
