+++ modules/objdetect/src/chessboard.cpp
@@ -4,6 +4,7 @@
 
 #include "precomp.hpp"
 #include "opencv2/flann.hpp"
+#include "opencv2/3d.hpp"
 #include "chessboard.hpp"
 #include "math.h"
