+++ modules/objdetect/src/precomp.hpp
@@ -48,6 +48,7 @@
 #include "opencv2/objdetect/barcode.hpp"
 #include "opencv2/imgproc.hpp"
 
+#include <opencv2/core/utils/logger.hpp>
 #include "opencv2/core/utility.hpp"
 #include "opencv2/core/ocl.hpp"
 #include "opencv2/core/private.hpp"
