+++ modules/objdetect/src/quadsubpix.cpp
