+++ modules/objdetect/test/test_precomp.hpp
@@ -5,6 +5,7 @@
 #define __OPENCV_TEST_PRECOMP_HPP__
 
 #include "opencv2/ts.hpp"
+#include "opencv2/3d.hpp"
 #include "opencv2/objdetect.hpp"
 
 #include <random>
